backend/requirements.txt
aiofiles==25.1.0
alembic==1.16.4
APScheduler==3.11.0
async-lru==2.0.5
bcrypt==5.0.0
DateTime==5.5
# Updated FastAPI version for Python 3.13 compatibility
fastapi[all]==0.116.2
loguru==0.7.3
openpyxl==3.1.5
pandas==2.3.2
Pillow==11.3.0
psutil==7.0.0
pydantic-validation-decorator==0.1.4
PyJWT[crypto]==2.10.1
PyMySQL==1.1.1
# Added aiomysql as an alternative async MySQL driver
aiomysql==0.2.0
redis==6.4.0
ruff==0.14.6
SQLAlchemy[asyncio]==2.0.43
sqlglot[rs]==27.8.0
user-agents==2.2.0

# OCR Dependencies
paddlepaddle>=2.6.0
paddleocr>=2.7.0.3
